[Fix #2] Initial reframing to rbhint
While RuboCop started as a tongue-in-cheek expression referencing a
movie from the 90s, the encroaching police state in America (and the
godawful reboot) make it less likely that casual references to RoboCop
will indicate the spirit of playfullness that the project was founded
in.

As a result, I am gently forking rubocop in an effort to perform the
work to reframe to something less enforcement-oriented and more oriented
towards gentle guidance towards community norms.

This is not a project I can do alone. I've got a near infinite set of
projects on my plate; and I would deeply appreciate the efforts of the
broader Ruby community to join me in doing this work.

In the words of Angela Davis:
> “You have to act as if it were possible to radically transform the
>  world. And you have to do it all the time.”

In the words of Emma Lazarus
> "Until we are all free, we are none of us free."