[#2, #15] Projects using `rbhint` still can use `rubocop` when necessary
Much of the tooling around code formatting assumes an executable named
`rubocop` available on the path. By addding this executable back in, we
make it slightly easier for folks to transition to `rbhint` without
breaking their development environment and without requiring tool
providers to change their code.

We may want to strongly indicate that the `rubocop` executable is
_deprecated_, but I'm not entirely sure how to go about doing that.