[#2] Projects using `rbhint` may run `rubocop` when necessary (#33)
Much of the tooling around code formatting assumes an executable named
`rubocop` available on the path. By addding this executable back in, we
make it slightly easier for folks to transition to `rbhint` without
breaking their development environment and without requiring tool
providers to change their code.

We may want to strongly indicate that the `rubocop` executable is
_deprecated_, but I'm not entirely sure how to go about doing that.

* [#2] Run CLI tests against the `rbhint` executable

This should ensure that the core use case (running `rbhint`) is checked
on CI and before relase.

It'd be nice if there was a way to test them _both_ without copy-pasting
but I don't have the bandwidth to work on that right now.